Twitter uploader (#22)
Adds support for a Twitter uploader to Go Note Go.

When the twitter uploader is selected, everything entered into go note go is tweeted.
Long text is split using a decent heuristic to get decent tweets.
Things entered in a single session form a single tweet-thread.
Press shift-enter to start a new session.

So if you submit text with shift-enter, it will show up as a new tweet. If you shift-enter with nothing entered, the next thing you type will be a new tweet.